+++ src/index.js
@@ -21,11 +21,6 @@ export { default as LayoutContainer } from "./components/LayoutContainer/LayoutC
 /**
  * buttons
  */
-export { ActionButton } from "./components/ActionButton/ActionButton";
-
-/**
- * buttons
- */
 export { Button } from "./components/Button/Button";
 
 /**
@@ -119,3 +114,13 @@ export { CheckBoxes } from "./components/CheckBoxes/CheckBoxes";
  * Check box
  */
 export { CheckBoxForm } from "./components/CheckBoxForm/CheckBoxForm";
+
+/*
+ * FormMultiTextField
+ */
+export { FormMultiTextField } from "./components/FormMultiTextField/FormMultiTextField";
+
+/*
+ * Review and Submit
+ */
+export { ReviewSubmit } from "./components/ReviewSubmit/ReviewSubmit";